Inventory Cost Displayer
Super small quick script that I threw together because I was bored. Similarly to how Hypixel Skyblock has their shop system set up, the script listens for an inventory being opened, then goes through the player's inventory and checks to see if any of the items have a cost flag, if they do then the appropriate lore is added, and if they don't the item isn't touched. The lore automatically goes back to what it was after the inventory that was opened is closed.

Feel free to build off of it/optimize it or whatever.